LayoutTests/fast/dom/shadow/content-element-distributed-nodes-expected.txt

Tests for content element's getDistributedNodes().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
Verify that getDistributedNodes() returns a static NodeList.
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.length is expectedNodes.length
PASS successfullyParsed is true

TEST COMPLETE
